requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
matplotlib==3.7.2
seaborn==0.12.2
yfinance==0.2.66
alpha-vantage==2.3.1
requests==2.31.0
nltk==3.8.1
textblob==0.17.1
streamlit==1.26.0
plotly==5.17.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
